# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Minut Booking API",
    "version": "1.0.0",
    "description": "REST API for managing rental units and guest reservations. All endpoints are versioned under /v1. Reservations use half-open date ranges [checkIn, checkOut) and overlapping reservations for the same unit are rejected with 409."
  },
  "servers": [
    {
      "url": "http://localhost:5006"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PageMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "ReservationList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Reservation"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "rentalUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "rentalUnit": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "name": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "name"
            ],
            "description": "Denormalised for list views"
          },
          "guestName": {
            "type": "string"
          },
          "guestEmail": {
            "type": [
              "string",
              "null"
            ]
          },
          "guestCount": {
            "type": "integer"
          },
          "notes": {
            "type": [
              "string",
              "null"
            ]
          },
          "checkIn": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2024-07-01"
          },
          "checkOut": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2024-07-01",
            "description": "Exclusive — the guest leaves this day"
          },
          "nights": {
            "type": "integer"
          },
          "version": {
            "type": "integer",
            "description": "Send back unchanged when updating"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "rentalUnitId",
          "rentalUnit",
          "guestName",
          "guestEmail",
          "guestCount",
          "notes",
          "checkIn",
          "checkOut",
          "nights",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "RevisionList": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Revision"
            }
          }
        },
        "required": [
          "data"
        ]
      },
      "Revision": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer"
          },
          "changeType": {
            "type": "string",
            "enum": [
              "update",
              "delete"
            ]
          },
          "snapshot": {
            "type": "object",
            "additionalProperties": {},
            "description": "The row as it was before the change"
          },
          "changedBy": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "changedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "version",
          "changeType",
          "snapshot",
          "changedBy",
          "changedAt"
        ]
      },
      "ReservationEnvelope": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Reservation"
          }
        },
        "required": [
          "data"
        ]
      },
      "CreateReservation": {
        "type": "object",
        "properties": {
          "rentalUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "guestName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Ada Lovelace"
          },
          "guestEmail": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 320,
            "format": "email",
            "example": "ada@example.com"
          },
          "guestCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50,
            "default": 1
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "checkIn": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2024-07-01"
          },
          "checkOut": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2024-07-01"
          }
        },
        "required": [
          "rentalUnitId",
          "guestName",
          "checkIn",
          "checkOut"
        ]
      },
      "UpdateReservation": {
        "type": "object",
        "properties": {
          "rentalUnitId": {
            "type": "string",
            "format": "uuid"
          },
          "guestName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "Ada Lovelace"
          },
          "guestEmail": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 320,
            "format": "email",
            "example": "ada@example.com"
          },
          "guestCount": {
            "type": "integer",
            "minimum": 1,
            "maximum": 50
          },
          "notes": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "checkIn": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2024-07-01"
          },
          "checkOut": {
            "type": "string",
            "format": "date",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2024-07-01"
          },
          "version": {
            "type": "integer",
            "minimum": 1,
            "description": "The version you last read. Mismatch → 409 VERSION_CONFLICT"
          }
        },
        "required": [
          "version"
        ]
      }
    },
...
  },
  "paths": {
...
    "/v1/rental-units/{id}/reservations": {
      "get": {
        "tags": [
          "Rental units"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List reservations of a rental unit",
        "description": "Optionally limited to those intersecting the half-open period [from, to).",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "description": "1-based page number"
            },
            "required": false,
            "description": "1-based page number",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "Items per page (max 100)"
            },
            "required": false,
            "description": "Items per page (max 100)",
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2024-07-01",
              "description": "Include stays ending after this date"
            },
            "required": false,
            "description": "Include stays ending after this date",
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2024-07-01",
              "description": "Include stays starting before this date"
            },
            "required": false,
            "description": "Include stays starting before this date",
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "checkIn",
                "-checkIn",
                "createdAt",
                "-createdAt"
              ],
              "default": "checkIn"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/reservations": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List reservations",
        "description": "Filter by rental unit and/or by time. A reservation matches a period when its half-open stay [checkIn, checkOut) intersects the half-open period [from, to).",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "description": "1-based page number"
            },
            "required": false,
            "description": "1-based page number",
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "description": "Items per page (max 100)"
            },
            "required": false,
            "description": "Items per page (max 100)",
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": false,
            "name": "rentalUnitId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2024-07-01",
              "description": "Include stays ending after this date"
            },
            "required": false,
            "description": "Include stays ending after this date",
            "name": "from",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
              "example": "2024-07-01",
              "description": "Include stays starting before this date"
            },
            "required": false,
            "description": "Include stays starting before this date",
            "name": "to",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "checkIn",
                "-checkIn",
                "createdAt",
                "-createdAt"
              ],
              "default": "checkIn"
            },
            "required": false,
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reservations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create a reservation",
        "description": "Requires admin. Returns 409 RESERVATION_OVERLAP if the unit is already booked for any of those nights.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (overlap, stale version, or dependent records)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reservations/{id}": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get a reservation",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Reservations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update a reservation",
        "description": "Partial update with optimistic locking via `version`. Date changes are re-checked for overlap.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReservation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict (overlap, stale version, or dependent records)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Reservations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Cancel a reservation",
        "description": "Soft delete; the nights become available again.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reservations/{id}/history": {
      "get": {
        "tags": [
          "Reservations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Change history of a reservation",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Resource id (UUID v4)"
            },
            "required": true,
            "description": "Resource id (UUID v4)",
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
...
}